Removed `Fragment` used for UI, using `Activity` instead. Handling orientation changes ourselves.
